- The beginnings of an undo/redo system, only currently works for various (not all) actions in the resource editor... probably not 100% safe to use yet :P
git-svn-id: http://mancubus.net/svn/hosted/slade/trunk@1358 1ddb354d-3e0c-0410-833a-d70222086f97

+++ src/GameConfiguration.cpp
@@ -371,11 +371,11 @@ bool GameConfiguration::mapFormatSupported(int map_format, int game, int port) {
 		return false;
 
 	// Check port if one specified
-	if (port >= 0 && port <= port_configs.size())
+	if (port >= 0 && port <= (int)port_configs.size())
 		return port_configs[port].supported_formats[map_format];
 
 	// Check game
-	if (game >= 0 && game <= game_configs.size())
+	if (game >= 0 && game <= (int)game_configs.size())
 		return game_configs[game].supported_formats[map_format];
 
 	// Not supported
